+++ controllers/rest/version/Version.go
@@ -1,4 +1,4 @@
-package controllers
+package version
 
 import (
 	"net/http"
