+++ controllers/ws/position/Postion.go
@@ -1,0 +1,52 @@
+package position
+
+import (
+	"context"
+
+	"github.com/BloomGameStudio/PlayerService/controllers/ws"
+	"github.com/labstack/echo/v4"
+)
+
+// NOTE: We may need to adjust default configuration and values
+// examples:
+// https://github.com/gorilla/websocket/blob/master/examples/command/main.go
+
+func Position(c echo.Context) error {
+
+	// QUESTION: Is this needed?
+	// Only changes will be sent the only exception to this is the opening/first request where the full state will be sent
+	// Partial player data can be received or full
+	// TODO: Partial Reads
+	// TODO: Partial Writes
+
+	ws, err := ws.Upgrader.Upgrade(c.Response(), c.Request(), nil)
+	if err != nil {
+		return err
+	}
+	defer ws.Close()
+
+	writerChan := make(chan error)
+	readerChan := make(chan error)
+
+	timeoutCTX, timeoutCTXCancel := context.WithCancel(context.Background())
+	defer timeoutCTXCancel()
+
+	go positionWriter(c, ws, writerChan, timeoutCTX)
+	go positionReader(c, ws, readerChan, timeoutCTX)
+
+	// QUESTION: Do we want to wait on both routines to error out?
+	// Return nil if either the reader or the writer encounters a error
+	// Do NOT return the error this will cause the error "the connection is hijacked"
+
+	select {
+
+	case w := <-writerChan:
+		c.Logger().Debugf("Recieved writerChan error: %v", w)
+		return nil
+
+	case r := <-readerChan:
+		c.Logger().Debugf("Recieved readerChan error: %v", r)
+		return nil
+
+	}
+}
