+++ controllers/ws/state/State.go
@@ -1,4 +1,4 @@
-package controllers
+package state
 
 // import (
 // 	"errors"
