+++ main.go
@@ -1,7 +1,13 @@
 package main
 
 import (
-	"github.com/BloomGameStudio/PlayerService/controllers"
+	"github.com/BloomGameStudio/PlayerService/controllers/rest/ping"
+	"github.com/BloomGameStudio/PlayerService/controllers/rest/player"
+	"github.com/BloomGameStudio/PlayerService/controllers/rest/version"
+	"github.com/BloomGameStudio/PlayerService/controllers/ws/hello"
+	wsPlayer "github.com/BloomGameStudio/PlayerService/controllers/ws/player"
+	"github.com/BloomGameStudio/PlayerService/controllers/ws/position"
+	"github.com/BloomGameStudio/PlayerService/controllers/ws/rotation"
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
 	"github.com/spf13/viper"
@@ -28,13 +34,13 @@ func main() {
 	// HTTP routes
 
 	// HTTP Testing routes
-	e.GET("ping", controllers.Ping)
-	e.GET("v", controllers.Version)
+	e.GET("ping", ping.Ping)
+	e.GET("v", version.Version)
 	// End of HTTP testing routes
 
 	// Player routes
-	e.GET("player", controllers.GetPlayer)
-	e.POST("player", controllers.CreatePlayer)
+	e.GET("player", player.GetPlayer)
+	e.POST("player", player.CreatePlayer)
 	// End Player routes
 
 	// End of HTTP routes
@@ -45,13 +51,13 @@ func main() {
 	// Web Socket Testing routes
 	ws.File("", "public/index.html") // http://127.0.0.1:1323/ws/
 	// ws://localhost:1323/ws
-	ws.GET("hello", controllers.Hello)
+	ws.GET("hello", hello.Hello)
 	// End of Web Socket testing routes
 
-	ws.GET("player", controllers.Player)
+	ws.GET("player", wsPlayer.Player)
 	// ws.GET("state", controllers.State)
-	ws.GET("position", controllers.Position)
-	ws.GET("rotation", controllers.Rotation)
+	ws.GET("position", position.Position)
+	ws.GET("rotation", rotation.Rotation)
 
 	port := viper.GetString("PORT")
 	e.Logger.Fatal(e.Start(":" + port))
